+++ code/renderercommon/tr_image_png.c
@@ -540,7 +540,7 @@ static uint32_t DecompressIDATs(struct BufferedFile *BF, uint8_t **Buffer)
 
 	if(!(BF && Buffer))
 	{
-		return(-1);
+		return((uint32_t)-1);
 	}
 
 	/*
@@ -561,7 +561,7 @@ static uint32_t DecompressIDATs(struct BufferedFile *BF, uint8_t **Buffer)
 
 	if(!FindChunk(BF, PNG_ChunkType_IDAT))
 	{
-		return(-1);
+		return((uint32_t)-1);
 	}
 
 	/*
@@ -584,7 +584,7 @@ static uint32_t DecompressIDATs(struct BufferedFile *BF, uint8_t **Buffer)
 
 			BufferedFileRewind(BF, BytesToRewind);
 
-			return(-1);
+			return((uint32_t)-1);
 		}
 
 		/*
@@ -621,7 +621,7 @@ static uint32_t DecompressIDATs(struct BufferedFile *BF, uint8_t **Buffer)
 			{
 				BufferedFileRewind(BF, BytesToRewind);
 
-				return(-1);
+				return((uint32_t)-1);
 			}
 
 			BytesToRewind += Length + PNG_ChunkCRC_Size;
@@ -634,7 +634,7 @@ static uint32_t DecompressIDATs(struct BufferedFile *BF, uint8_t **Buffer)
 	CompressedData = ri.Malloc(CompressedDataLength);
 	if(!CompressedData)
 	{
-		return(-1);
+		return((uint32_t)-1);
 	}
 
 	CompressedDataPtr = CompressedData;
@@ -654,7 +654,7 @@ static uint32_t DecompressIDATs(struct BufferedFile *BF, uint8_t **Buffer)
 		{
 			ri.Free(CompressedData); 
 
-			return(-1);
+			return((uint32_t)-1);
 		}
 
 		/*
@@ -688,14 +688,14 @@ static uint32_t DecompressIDATs(struct BufferedFile *BF, uint8_t **Buffer)
 			{
 				ri.Free(CompressedData); 
 
-				return(-1);
+				return((uint32_t)-1);
 			}
 
 			if(!BufferedFileSkip(BF, PNG_ChunkCRC_Size))
 			{
 				ri.Free(CompressedData); 
 
-				return(-1);
+				return((uint32_t)-1);
 			}
 
 			memcpy(CompressedDataPtr, OrigCompressedData, Length);
@@ -726,7 +726,7 @@ static uint32_t DecompressIDATs(struct BufferedFile *BF, uint8_t **Buffer)
 	{
 		ri.Free(CompressedData);
 
-		return(-1);
+		return((uint32_t)-1);
 	}
 
 	/*
@@ -738,7 +738,7 @@ static uint32_t DecompressIDATs(struct BufferedFile *BF, uint8_t **Buffer)
 	{
 		ri.Free(CompressedData);
 
-		return(-1);
+		return((uint32_t)-1);
 	}
 
 	/*
@@ -769,7 +769,7 @@ static uint32_t DecompressIDATs(struct BufferedFile *BF, uint8_t **Buffer)
 	{
 		ri.Free(DecompressedData);
 
-		return(-1);
+		return((uint32_t)-1);
 	}
 
 	/*
@@ -2372,7 +2372,7 @@ void R_LoadPNG(const char *name, byte **pic, int *width, int *height)
 	 *  Rewind to the start of the file.
 	 */
 
-	if(!BufferedFileRewind(ThePNG, -1))
+	if(!BufferedFileRewind(ThePNG, (unsigned)-1))
 	{
 		CloseBufferedFile(ThePNG);
